+++ app/Models/CourseLessonVideo.php
@@ -7,31 +7,27 @@ use App\Traits\Uuid;
 use Spatie\Sluggable\HasSlug;
 use Spatie\Sluggable\SlugOptions;
 
-class CourseLessonVideo extends Model
-{
+class CourseLessonVideo extends Model {
     use Uuid, HasSlug;
 
     protected $fillable = ['name', 'slug', 'description', 'course_lesson_id', 'video', 'order'];
     protected $hidden = ['id', 'course_lesson_id', 'video', 'created_at', 'updated_at'];
 
-    public function getSlugOptions(): SlugOptions
-    {
+    public function getSlugOptions(): SlugOptions {
         return SlugOptions::create()
             ->generateSlugsFrom('name')
             ->saveSlugsTo('slug');
     }
 
-    public function lesson()
-    {
+    public function lesson() {
         return $this->belongsTo(CourseLesson::class, 'course_lesson_id');
     }
 
-    public static function booted()
-    {
+    public static function booted() {
         parent::boot();
 
-        self::deleting(function ($model) {
-            if (file_exists(storage_path('app/private/' . $model->video))) {
+        self::deleted(function ($model) {
+            if (isset($model->video) && file_exists(storage_path('app/private/' . $model->video))) {
                 unlink(storage_path('app/private/' . $model->video));
             }
         });
